Rename tasks.py to factory.py
