+++ python/src/tests/tasks/test_factory.py
@@ -1,10 +1,8 @@
 import pytest
-import os
-from tasks.tasks import TaskFactory
+from tasks.factory import TaskFactory
 from result import Result
 from config import get_config
 from mock import Mock, patch
-from helpers.count_matrix import CountMatrix
 
 config = get_config()
 
@@ -12,7 +10,7 @@ config = get_config()
 class TestTaskFactory:
     @pytest.fixture(autouse=True)
     def set_mock_task_factory(self):
-        with patch("tasks.tasks.CountMatrix") as MockCountMatrix:
+        with patch("tasks.factory.CountMatrix") as MockCountMatrix:
             instance = MockCountMatrix.return_value
             instance.sync.return_value = Mock()
             self.task_factory = TaskFactory()
